requirements.txt
streamlit>=1.28.0
python-dotenv==1.0.0
requests==2.31.0
groq>=0.4.1
pydantic>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
streamlit-google-auth>=0.1.0
reportlab>=4.0.0
flask>=2.3.0
pydeck>=0.8.0
pytest>=7.0.0